refactor(openai): extract Responses-API content guards and file uploads
Continue decomposing ModelHandlerOpenAIResponse by extracting two more
cohesive concerns into stateless modules, following the same pattern as the
transport/streaming split. Behavior-preserving.

- openAIResponseContent.ts: pure content guards and value-object builders
  (createInputText, isMessageItem, isAssistantTextMessage,
  extractTextContentPart, isOutputMessage, isInputFileContent), shared by
  the handler and the file-upload helpers.
- openAIResponseFileUploads.ts: file/attachment upload infrastructure
  (uploadInlineInputFiles, uploadToolAttachments, buildInlineAttachmentParts,
  INLINEABLE_FILE_MIME_TYPES, UploadedOpenAIResponseAttachment). The handler
  now passes a logger and an OpenRouter-routing flag instead of relying on
  instance state.

Also drops imports that became dead after the extraction (toFile,
APIConnectionTimeoutError, OpenAIAPIError, OFFICE_MIME_TYPES,
loadAttachmentBuffer, isNonEmptyString, Buffer, path, and several response
types). Handler drops from 2839 to 2545 lines (3330 before the refactor).
